hinzufuegen einer optischen darstellung der Spielfeldbegrenzung. Rote lineie = ende des Spielfeldes. Gelbe linie = nahes ende des Spielfeldes. aenderung des gameplays: uebertritt ein spieler die Rote line wird er um das hundertfache langsammer und erhaellt kontinuirlichen schaden.

+++ src/player.h
@@ -18,7 +18,7 @@ class Player : public movableObject
     Ogre::String getObjektName();
 //    int getNumber();
     bool BoostAvaible();
-    void takeDamage(int arg1);
+    void takeDamage(float arg1);
     
         
     // wenn tot dann kann er nicht mehr fliegen...
